Added support for Repeat Mode's to Image
I'm seeing scaling/rounding issues when using patterns to draw images.  Most sprites don't need texture wrapping turned on, so we should just use drawImage when no repeat is desired. 

I was seeing a lot of artifacts on Safari and what I believe to be rounding issues due to hierarchical concatenation of transforms.

Not sure if you are interested in this change, but I think it's pretty useful.